+++ src/constants/StateKeys.ts
@@ -1,3 +1,4 @@
 export const StateKeys = {
   executingDemoFile: "executingDemoFile",
+  variables: "variables",
 };