+++ src/extension.ts
@@ -10,9 +10,13 @@ import { PresenterView } from "./presenterView/PresenterView";
 import { NotesService } from "./services/NotesService";
 import { DemoApi } from "./services/DemoApi";
 import { UriHandler } from "./services/UriHandler";
+import { clearVariablesState } from "./utils";
 
 export async function activate(context: vscode.ExtensionContext) {
   Extension.getInstance(context);
+
+  // Clearing the variable state when the extension starts
+  clearVariablesState();
 
   DecoratorService.register();